Debug

+++ .github/workflows/full_test_suite/solana_init.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+if [ "$(whoami)" != "root" ]
+then
+    sudo su -s "$0"
+    exit
+fi
 
 # Install docker
 apt-get remove docker docker-engine docker.io containerd runc
